refactor: rename project to 'skillsync' and restructure components for better organization
