+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "nextproject",
+  "name": "skillsync",
   "version": "0.1.0",
   "private": true,
   "scripts": {
