+++ src/app/templates/page.jsx
@@ -1,4 +1,4 @@
-import Navbar from '../components/Navbar'
+import Navbar from '@/component/Navbar'
 import { FiArrowRight } from 'react-icons/fi'
 
 const templates = [
